@@ -22,30 +22,28 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@angular/common":  "~2.0.1",
-    "@angular/compiler":  "~2.0.1",
-    "@angular/core":  "~2.0.1",
-    "@angular/forms": "~2.0.1",
-    "@angular/http":  "~2.0.1",
-    "@angular/platform-browser":  "~2.0.1",
-    "@angular/platform-browser-dynamic":  "~2.0.1",
-    "@angular/router":  "~3.0.1",
-    "@angular/upgrade":  "~2.0.1",
-
-    "angular-in-memory-web-api": "~0.1.1",
+    "@angular/common": "~5.0.0",
+    "@angular/compiler": "~5.0.0",
+    "@angular/core": "~5.0.0",
+    "@angular/forms": "~5.0.0",
+    "@angular/http": "~5.0.0",
+    "@angular/platform-browser": "~5.0.0",
+    "@angular/platform-browser-dynamic": "~5.0.0",
+    "@angular/router": "~5.0.0",
+    "@angular/upgrade": "~5.0.0",
+    "angular-in-memory-web-api": "~0.4.0",
     "bootstrap": "^3.3.7",
     "systemjs": "0.19.39",
     "core-js": "^2.4.1",
     "reflect-metadata": "^0.1.8",
-    "rxjs": "5.0.0-beta.12",
+    "rxjs": "^5.5.0",
     "zone.js": "^0.6.25"
   },
   "devDependencies": {
     "concurrently": "^2.2.0",
     "lite-server": "^2.2.0",
-    "typescript": "^1.8.10",
+    "typescript": "^2.5.0",
     "typings": "^1.0.4",
-
     "canonical-path": "0.0.2",
     "http-server": "^0.9.0",
     "tslint": "^3.7.4",
