@@ -1,24 +1,23 @@
 import { NgModule }       from '@angular/core';
 import { BrowserModule }  from '@angular/platform-browser';
 import { FormsModule }    from '@angular/forms';
-import { HttpModule, JsonpModule } from '@angular/http';
+import { HttpClientModule, HttpClientJsonpModule } from '@angular/common/http';
 
-import { CommentBoxComponent } from './components/comment-box.component';
-import { CommentListComponent } from './components/comment-list.component';
-import { CommentFormComponent } from './components/comment-form.component';
-import { CommentComponent } from './components/index';
+import { CommentBoxComponent } from './Components/comment-box.component';
+import { CommentListComponent } from './Components/comment-list.component';
+import { CommentFormComponent } from './Components/comment-form.component';
+import { CommentComponent } from './Components/index';
 
 
-import { CommentService } from './services/comment.service';
+import { CommentService } from './Services/comment.service';
 
 
 @NgModule({
   imports: [
     BrowserModule,
     FormsModule,
-     HttpModule,
-    JsonpModule,
-    
+    HttpClientModule,
+    HttpClientJsonpModule
   ],
   declarations: [
     CommentBoxComponent,
