@@ -1,8 +1,8 @@
 /* * * ./app/comments/services/comment.service.ts * * */
 // Imports
 import { Injectable }     from '@angular/core';
-import { Http, Response, Headers, RequestOptions } from '@angular/http';
-import { Comment }           from '../model/comment';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { Comment }           from '../Model/comment';
 import {Observable} from 'rxjs/Rx';
 
 // Import RxJs required methods
@@ -11,48 +11,45 @@ import 'rxjs/add/operator/catch';
 
 @Injectable()
 export class CommentService {
-     // Resolve HTTP using the constructor
-     constructor (private http: Http) {}
-     // private instance variable to hold base url
-     private commentsUrl = 'http://localhost:3000/api/comments';
-    // private commentsUrl = 'http://578f454de2fa491100415d08.mockapi.io/api/Comment'; 
-     
+    // private instance variable to hold base url
+    private commentsUrl = 'http://localhost:3000/api/comments';
+    // private commentsUrl = 'http://578f454de2fa491100415d08.mockapi.io/api/Comment';
+    // Resolve HTTP using the constructor
+     constructor (private http: HttpClient) {}
+
      // Fetch all existing comments
      getComments() : Observable<Comment[]>{
          // ...using get request
          return this.http.get(this.commentsUrl)
-                        // ...and calling .json() on the response to return data
-                         .map((res:Response) => res.json())
-                         //...errors if any
+                        // ...now we return data
+                         .map(res => res )
+                         // ...errors if any
                          .catch((error:any) => Observable.throw(error.json().error || 'Server error'));
-        
      }
 
      // Add a new comment
     addComment (body: Object): Observable<Comment[]> {
         let bodyString = JSON.stringify(body); // Stringify payload
-        let headers = new Headers({ 'Content-Type': 'application/json' }); // ... Set content type to JSON
-        let options = new RequestOptions({ headers: headers }); // Create a request option
-
-        return this.http.post(this.commentsUrl, body, options) // ...using post request
-                         .map((res:Response) => res.json()) // ...and calling .json() on the response to return data
-                         .catch((error:any) => Observable.throw(error.json().error || 'Server error')); //...errors if any
-    }   
+        return this.http.post(this.commentsUrl, bodyString, {
+            headers: new HttpHeaders().set('Content-Type', 'application/json'), // adding headers to request
+        }) // ...using post request
+                         .map(res => res ) // ...now we return data
+                         .catch((error:any) => Observable.throw(error.json().error || 'Server error')); // ...errors if any
+    }
 
     // Update a comment
     updateComment (body: Object): Observable<Comment[]> {
         let bodyString = JSON.stringify(body); // Stringify payload
-        let headers = new Headers({ 'Content-Type': 'application/json' }); // ... Set content type to JSON
-        let options = new RequestOptions({ headers: headers }); // Create a request option
-
-        return this.http.put(`${this.commentsUrl}/${body['id']}`, body, options) // ...using put request
-                         .map((res:Response) => res.json()) // ...and calling .json() on the response to return data
-                         .catch((error:any) => Observable.throw(error.json().error || 'Server error')); //...errors if any
-    }   
+        return this.http.put(`${this.commentsUrl}/${body['id']}`, bodyString ,{
+            headers: new HttpHeaders().set('Content-Type', 'application/json'), // adding headers to request            
+        }) // ...using put request
+                         .map(res => res ) // ...now we return data
+                         .catch((error:any) => Observable.throw(error.json().error || 'Server error')); // ...errors if any
+    }
     // Delete a comment
     removeComment (id:string): Observable<Comment[]> {
         return this.http.delete(`${this.commentsUrl}/${id}`) // ...using put request
-                         .map((res:Response) => res.json()) // ...and calling .json() on the response to return data
-                         .catch((error:any) => Observable.throw(error.json().error || 'Server error')); //...errors if any
-    }   
+                         .map(res => res ) // ...now we return data
+                         .catch((error:any) => Observable.throw(error.json().error || 'Server error')); // ...errors if any
+    }
 }
